+++ apps/backend/src/modules/friends/friend.service.ts
@@ -103,6 +103,12 @@ class FriendService {
 			});
 		}
 
+		await this.notificationService.deleteAllNotificationsByParameters(
+			id,
+			userId,
+			NotificationType.NEW_FOLLOWER,
+		);
+
 		return isDeleted;
 	}
 
@@ -139,15 +145,18 @@ class FriendService {
 		count,
 		id,
 		page,
+		search,
 	}: {
 		count: number;
 		id: number;
 		page: number;
+		search: string;
 	}): Promise<PaginationResponseDto<UserAuthResponseDto>> {
 		const { items, total } = await this.friendRepository.getPotentialFollowers({
 			count,
 			id,
 			page: convertPageToZeroIndexed(page),
+			search,
 		});
 
 		return {
@@ -160,15 +169,18 @@ class FriendService {
 		count,
 		id,
 		page,
+		search,
 	}: {
 		count: number;
 		id: number;
 		page: number;
+		search: string;
 	}): Promise<PaginationResponseDto<UserAuthResponseDto>> {
 		const { items, total } = await this.friendRepository.getUserFollowers({
 			count,
 			id,
 			page: convertPageToZeroIndexed(page),
+			search,
 		});
 
 		return {
@@ -181,15 +193,18 @@ class FriendService {
 		count,
 		id,
 		page,
+		search,
 	}: {
 		count: number;
 		id: number;
 		page: number;
+		search: string;
 	}): Promise<PaginationResponseDto<UserAuthResponseDto>> {
 		const { items, total } = await this.friendRepository.getUserFollowings({
 			count,
 			id,
 			page: convertPageToZeroIndexed(page),
+			search,
 		});
 
 		return {