+++ apps/frontend/src/modules/friends/libs/types/types.ts
@@ -1,5 +1,6 @@
 export {
 	type FriendFollowRequestDto,
 	type FriendFollowResponseDto,
+	type FriendListRequestQueryDto,
 	type FriendUnfollowRequestDto,
 } from "@trackmates/shared";