+++ apps/frontend/src/pages/management/libs/components/groups-tab/libs/enums/groups-table-header.enum.ts
@@ -1,5 +1,5 @@
 const GroupsTableHeader = {
-	BUTTONS: "",
+	ACTIONS: "Actions",
 	ID: "ID",
 	NAME: "Group name",
 	PERMISSIONS: "Permissions",