+++ packages/shared/src/modules/courses/libs/enums/courses-api-path.enum.ts
@@ -1,5 +1,7 @@
 const CoursesApiPath = {
 	$COURSE_ID: "/:courseId",
+	FROM_VENDORS: "/from-vendors",
+	FROM_VENDORS_$COURSE_ID: "/from-vendors/:courseId",
 	RECOMMENDED: "/recommended",
 	ROOT: "/",
 } as const;